// src/redux-devtools-extension.d.ts
export interface Config extends EnhancerOptions {
	type?: string;
}

export interface StartAction {
	readonly type: "START";
	readonly state: undefined;
	readonly id: undefined;
	readonly source: typeof source;
}

export interface StopAction {
	readonly type: "STOP";
	readonly state: undefined;
	readonly id: undefined;
	readonly source: typeof source;
	readonly failed?: boolean;
}

export interface DispatchAction {
	readonly type: "DISPATCH";
	readonly payload: AppDispatchAction;
	readonly state: string | undefined;
	readonly id: string;
	readonly source: typeof source;
}

export interface ImportAction {
	readonly type: "IMPORT";
	readonly payload: undefined;
	readonly state: string | undefined;
	readonly id: string;
	readonly source: typeof source;
}

export interface ActionAction {
	readonly type: "ACTION";
	readonly payload: string | CustomAction;
	readonly state: string | undefined;
	readonly id: string;
	readonly source: typeof source;
}

export interface ExportAction {
	readonly type: "EXPORT";
	readonly payload: undefined;
	readonly state: string | undefined;
	readonly id: string;
	readonly source: typeof source;
}

export interface UpdateAction {
	readonly type: "UPDATE";
	readonly state: string | undefined;
	readonly id: string;
	readonly source: typeof source;
}

export interface OptionsAction {
	readonly type: "OPTIONS";
	readonly options: Options;
	readonly id: undefined;
	readonly source: typeof source;
}

export interface ImportStatePayload<S, A extends Action<string>> {
	readonly type: "IMPORT_STATE";
	readonly nextLiftedState: LiftedState<S, A, unknown> | readonly A[];
	readonly preloadedState?: S;
}

export interface ImportStateDispatchAction<S, A extends Action<string>> {
	readonly type: "DISPATCH";
	readonly payload: ImportStatePayload<S, A>;
}

export type ListenerMessage<S, A extends Action<string>> =
	| StartAction
	| StopAction
	| DispatchAction
	| ImportAction
	| ActionAction
	| ExportAction
	| UpdateAction
	| OptionsAction
	| ImportStateDispatchAction<S, A>;

export interface ConnectResponse {
	init: <S>(state: S) => void;
	subscribe: <S, A extends Action<string>>(
		listener: (message: ListenerMessage<S, A>) => void,
	) => (() => void) | undefined;
	unsubscribe: () => void;
	send: <S, A extends Action<string>>(action: A, state: S) => void;
	error: (payload: string) => void;
}

export interface ReduxDevtoolsExtension {
	connect: (preConfig: Config) => ConnectResponse;
}